+++ src/bmlite/P2D/domains.py
@@ -188,9 +188,9 @@ class Electrode:
             Keyword arguments to set the electrode attributes. The required
             keys and descriptions are given below:
 
-            ========== ========================================================
+            ========== =======================================================
             Key        Description [units] (type)
-            ========== ========================================================
+            ========== =======================================================
             Nx         number of `x` discretizations [-] (*int*)
             Nr         number of `r` discretizations [-] (*int*)
             thick      electrode thickness [m] (*float*)
@@ -207,8 +207,9 @@ class Electrode:
             i0_deg     `i0` degradation factor [-] (*float*)
             Ds_deg     `Ds` degradation factor [-] (*float*)
             material   class name from `bmlite.materials` [-] (*str*)
+            csvfile    path to CSV file with OCV data (optional) (*str*)
             submodels  `submodels` classes to include (*dict[dict]*)
-            ========== ========================================================
+            ========== =======================================================
 
         """
         from . import submodels
@@ -234,6 +235,7 @@ class Electrode:
         self.i0_deg = kwargs.get('i0_deg')
         self.Ds_deg = kwargs.get('Ds_deg')
         self.material = kwargs.get('material')
+        self.csvfile = kwargs.get('csvfile', None)
 
         self.update()
 
@@ -258,6 +260,7 @@ class Electrode:
             `A_s = 3 * eps_AM / R_s`
 
         """
+        import inspect
         from .. import materials
 
         self.eps_void = 1. - self.eps_s - self.eps_el
@@ -274,7 +277,12 @@ class Electrode:
             raise ValueError('eps_s <= eps_CBD.')
 
         Material = getattr(materials, self.material)
-        self._material = Material(self.alpha_a, self.alpha_c, self.Li_max)
+
+        if 'csvfile' in inspect.signature(Material).parameters:
+            self._material = Material(self.alpha_a, self.alpha_c, self.Li_max,
+                                      csvfile=self.csvfile)
+        else:
+            self._material = Material(self.alpha_a, self.alpha_c, self.Li_max)
 
     def get_Ds(self, x: float | np.ndarray, T: float,
                fluxdir: float | np.ndarray) -> float | np.ndarray: